Change the extension to only display the icon (no text) on the toolbar
Added another contribution to have different behaviour in the menus and on
the toolbar. By adding another target to the existing contribution I could
not achieve that.

I also had to change the target of the existing contribution to avoid
having two buttons (one with text) on the toolbar. This also means that the
menu option does not appear when right clicking in the query results list.
There it would be necessary to have yet another way of retrieving the id
of the work item.